Created Object.h file to hold object class, functions.h file for functions class header,  and functions.cpp for the functions
