+++ Object.h
@@ -1,0 +1,19 @@
+#ifndef OBJECT_H
+#define OBJECT_H
+
+#include <string>
+
+using namespace std;
+
+// Obj struct
+struct Object {
+    string sequence;
+    int x;
+    int y;
+    int centerX;
+    int centerY;
+    int size;
+    char id;
+};
+
+#endif // OBJECT_H
