+++ project3.html
@@ -77,7 +77,7 @@
     
     <div class="container">
         <h1>Team Project - Hangman</h1>
-        <form enctype="multipart/form-data" action="../cgi-bin/project" method="post">
+        <form enctype="multipart/form-data" action="../../cgi-bin/project3" method="post">
             <br />
             
             <label for="name">Enter a random string: </label>
